+++ examples/distributed/all_reduce.py
@@ -25,7 +25,7 @@ import helion
 from helion._testing import DEVICE
 from helion._testing import run_example
 import helion.language as hl
-from helion.runtime.dist_utils import symm_mem_sync
+from helion.runtime.triton.dist_utils import symm_mem_sync
 
 # %%
 # Work around before symm mem natively supports extract dev_ptrs as tensors: from_blob
