+++ examples/distributed/allreduce_bias_rmsnorm.py
@@ -23,7 +23,7 @@ import helion
 from helion._testing import DEVICE
 from helion._testing import run_example
 import helion.language as hl
-from helion.runtime.dist_utils import symm_mem_sync
+from helion.runtime.triton.dist_utils import symm_mem_sync
 
 
 @helion.jit(
