+++ helion/_compiler/triton/backend.py
@@ -385,7 +385,7 @@ class TritonBackend(Backend):
             "triton_helpers": "from torch._inductor.runtime import triton_helpers",
             "tl_math": "from torch._inductor.runtime.triton_helpers import math as tl_math",
             "libdevice": "from torch._inductor.runtime.triton_compat import libdevice",
-            "helion_dist_utils": "from helion.runtime import dist_utils as helion_dist_utils",
+            "helion_dist_utils": "from helion.runtime.triton import dist_utils as helion_dist_utils",
             "nvshmem": "import torch.distributed._symmetric_memory._nvshmem_triton as nvshmem",
             "requires_nvshmem": "from torch.distributed._symmetric_memory._nvshmem_triton import requires_nvshmem",
             "_default_launcher": "from helion.runtime import default_launcher as _default_launcher",
